parse materials
